ai_models/requirements.txt
# AI Model Dependencies
openai>=1.0.0
openai-whisper>=20230918
numpy>=1.20.0
torch>=2.0.0
spacy>=3.5.0
textstat>=0.7.3
scikit-learn>=1.0.0
scipy>=1.8.0
pandas>=1.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Audio Processing
pyaudio>=0.2.13
ffmpeg-python>=0.2.0

# Testing and Development
pytest>=7.0.0
asyncio>=3.4.3
aiohttp>=3.8.0

# You can install the required spaCy models with:
# python -m spacy download en_core_web_md 